+++ Changes.md
@@ -17,6 +17,8 @@ See the Mod Description or [README](https://gitlab.com/CDAGaming/CraftPresence) 
     * Classgraph (`4.8.162` -> `4.8.163`)
     * JUnixSocket (`2.7.0` -> `2.8.1`)
 * Adjusted Mod Initialization in Forge 1.13+ to better detect (and avoid) running the mod on server-side
+* Adjusted the `general.brand` placeholder to support the `minecraft.launcher.brand` System Property
+    * The prior implementation of this placeholder will be used if this property is not present
 * Updated Translations for Belarusian, Pirate English, French, and Russian
     * Please note these are community-created translations from Crowdin, and issues may be present!